🐛 fix: change execa with spawn of node built-in child_process moudule due to module import error
